A2 Missile, Powerup, MainMenu, UI
Added missile powerup to game.
Powerup spawns randomly between 3 and 10 seconds.
Missile fired by RMB.
Added SFX to missile launch.
Added background music to MainMenu.
Added powerup  for missiles which triggers when it collides with the
player.
Powerup plays SFX when collected.
Powerup is destroyed when it hits the trigger for the death plane.
Player keeps track of missiles collected and will not shoot missiles
when it has 0.
Player also updates the UI with the amount of missiles it currently has.
Built .exe file for project.