Bex Bullet and Platforms
Added functionality to the bullets.
Bullets now spawn at the end of the gun and travel in the direction
sprite is facing.
Bullets use up their lifetime and then are destroyed.
Added fireRate to BexPlayer.
Imported spritesheet for metal platforms.
Setup initial platforms.
Velocity is used for movements now instead of AddForce.